refactor(problem-read): 문제 조회 시, api 호출하여 course 조회 하도록 변경
