+++ app/api/mathrank-problem-read-api/build.gradle
@@ -7,5 +7,6 @@ dependencies {
     implementation project(':app:api:mathrank-api-common')
     implementation project(':domain:mathrank-problem-domain')
     implementation project(':client:internal:mathrank-member-client')
+    implementation project(':client:internal:mathrank-course-client')
     implementation project(':client:external:mathrank-school')
 }
